WifiScanningService: always allocate enough buckets
On some devices, mSettings.buckets may be allocated before
WifiNative.getScanCapabilities is invoked. Since the size of
mSettings.buckets is determined by a value obtained by getting
the scan capabilities, it's size would be set to zero. This leads
to array bounds errors later in the code that assume the
array is at least as long as mTimeBuckets.

This patch makes sure that the size of mSettings.buckets is at
least as long as mTimeBuckets, so that array bounds errors are
avoided.

Thanks to Benjamin Legrand (scritch007) for his investigation of
the issue and original patch that led me in this direction.

Change-Id: Ic3fffd82c745922bd6378062ee188841880895a2